backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.4.0
python-dotenv==1.0.1
google-ai-generativelanguage==0.6.9
google-generativeai==0.8.1
